include key images of spent outputs when txs created

+++ src/daemon/monero_daemon_model.cpp
@@ -379,7 +379,7 @@ namespace monero {
         tgt->m_outputs.push_back(output_copy);
       }
     }
-    if (!src->m_output_indices.empty()) tgt->m_output_indices = std::vector<uint32_t>(src->m_output_indices);
+    if (!src->m_output_indices.empty()) tgt->m_output_indices = std::vector<uint64_t>(src->m_output_indices);
     tgt->m_metadata = src->m_metadata;
     tgt->m_common_tx_sets = src->m_common_tx_sets;
     if (!src->m_extra.empty()) throw std::runtime_error("extra deep copy not implemented");  // TODO: implement extra
